backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
firebase-admin==6.5.0
rdkit==2024.3.1
requests==2.31.0
python-jose==3.3.0
httpx==0.27.0
pydantic==2.7.1
python-multipart==0.0.9
